feat: Search 화면 검색 기록 로그 등록 추가
- Search 화면의 검색 내역 저장 추가

Related to: #13
