+++ app/src/main/java/com/example/findhomes/domain/repository/SearchRepository.kt
@@ -31,4 +31,7 @@ interface SearchRepository {
 
     suspend fun getStatisticsData(
     ) : List<SearchStatisticsResponse>?
+
+    suspend fun postSearchLogsData(
+    ) : String?
 }
