+++ app/src/main/java/com/example/findhomes/presentation/ui/search/SearchViewModel.kt
@@ -19,6 +19,7 @@ import com.example.findhomes.domain.usecase.search.GetSearchStatisticsUseCase
 import com.example.findhomes.domain.usecase.search.PostChatDataUseCase
 import com.example.findhomes.domain.usecase.search.PostManConUseCase
 import com.example.findhomes.domain.usecase.search.PostSearchFavoriteUseCase
+import com.example.findhomes.domain.usecase.search.PostSearchLogsDataUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -30,7 +31,8 @@ class SearchViewModel @Inject constructor(
     private val postManConUseCase: PostManConUseCase,
     private val getSearchDetailDataUseCase: GetSearchDetailDataUseCase,
     private val getSearchStatisticsUseCase: GetSearchStatisticsUseCase,
-    private val postSearchFavoriteUseCase: PostSearchFavoriteUseCase
+    private val postSearchFavoriteUseCase: PostSearchFavoriteUseCase,
+    private val postSearchLogsDataUseCase: PostSearchLogsDataUseCase
 ): ViewModel() {
     private val _searchData = MutableLiveData<List<SearchCompleteResponse>?>()
     val searchData: LiveData<List<SearchCompleteResponse>?> = _searchData
@@ -125,6 +127,17 @@ class SearchViewModel @Inject constructor(
                 Log.d("SearchViewModel", "로드된 데이터: $data")
             } catch (e: Exception) {
                 Log.e("SearchViewModel", "loadSearchStatisticsData 오류", e)
+            }
+        }
+    }
+
+    fun postSearchLogsData(){
+        viewModelScope.launch {
+            try{
+                postSearchLogsDataUseCase()
+                Log.d("SearchViewModel", "로그 잘 보내짐")
+            } catch (e: Exception) {
+                Log.e("SearchViewModel", "postSearchLogsData 오류", e)
             }
         }
     }
