+++ app/src/main/res/layout/fragment_home.xml
@@ -52,7 +52,8 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="20dp"
         android:padding="20dp"
-        android:background="@drawable/rectangle_green_8"
+        android:background="@drawable/rectangle_gray_8"
+        android:elevation="5dp"
         app:layout_constraintStart_toStartOf="@id/gl_start"
         app:layout_constraintTop_toBottomOf="@id/home_iv_title"
         app:layout_constraintBottom_toBottomOf="@id/home_iv_title"
@@ -91,86 +92,81 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <LinearLayout
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/btn_wish"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"
         app:layout_constraintStart_toStartOf="@id/gl_start"
         app:layout_constraintEnd_toEndOf="@id/gl_end"
-        android:layout_marginTop="10dp"
-        app:layout_constraintTop_toBottomOf="@id/btn_search">
+        app:layout_constraintTop_toBottomOf="@id/btn_search"
+        android:layout_height="wrap_content"
+        android:elevation="5dp"
+        android:background="@drawable/rectangle_gray_8"
+        android:padding="10dp"
+        android:layout_marginTop="20dp">
 
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/btn_wish"
-            android:layout_width="0dp"
+        <TextView
+            android:id="@+id/tv_home_wish_title"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:background="@drawable/rectangle_gray_8"
-            android:padding="10dp"
-            android:layout_marginEnd="5dp"
-            android:layout_marginTop="10dp">
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:text="찜 목록"
+            android:fontFamily="@font/pretendard_semibold"
+            android:textSize="20sp"
+            app:layout_constraintTop_toTopOf="parent" />
 
-            <TextView
-                android:id="@+id/tv_home_wish_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                android:text="찜 목록"
-                android:fontFamily="@font/pretendard_semibold"
-                android:textSize="20sp"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:id="@+id/tv_home_wish_sub"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:layout_marginTop="10dp"
-                android:textSize="14sp"
-                android:fontFamily="@font/pretendard_regular"
-                android:text="등록하셨던 관심 매물들을\n볼 수 있어요"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/tv_home_wish_title"/>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/btn_my"
-            android:layout_width="0dp"
+        <TextView
+            android:id="@+id/tv_home_wish_sub"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:background="@drawable/rectangle_gray_8"
-            android:padding="10dp"
-            android:layout_marginStart="5dp"
-            android:layout_marginTop="10dp">
+            android:gravity="center"
+            android:layout_marginTop="10dp"
+            android:textSize="14sp"
+            android:fontFamily="@font/pretendard_regular"
+            android:text="등록하셨던 관심 매물들을\n볼 수 있어요"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/tv_home_wish_title"/>
 
-            <TextView
-                android:id="@+id/tv_home_my_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                android:fontFamily="@font/pretendard_semibold"
-                android:textSize="20sp"
-                android:text="마이페이지"
-                app:layout_constraintTop_toTopOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
-            <TextView
-                android:id="@+id/tv_home_my_sub"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:layout_marginTop="10dp"
-                android:textSize="14sp"
-                android:fontFamily="@font/pretendard_regular"
-                android:text="사용자님의 등록했던 정보를\n확인할 수 있어요"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/tv_home_my_title"/>
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/btn_my"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:background="@drawable/rectangle_gray_8"
+        app:layout_constraintStart_toStartOf="@id/gl_start"
+        app:layout_constraintEnd_toEndOf="@id/gl_end"
+        app:layout_constraintTop_toBottomOf="@id/btn_wish"
+        android:elevation="5dp"
+        android:layout_marginTop="20dp"
+        android:padding="10dp">
 
-        </androidx.constraintlayout.widget.ConstraintLayout>
+        <TextView
+            android:id="@+id/tv_home_my_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:fontFamily="@font/pretendard_semibold"
+            android:textSize="20sp"
+            android:text="마이페이지"
+            app:layout_constraintTop_toTopOf="parent" />
 
-    </LinearLayout>
+        <TextView
+            android:id="@+id/tv_home_my_sub"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:layout_marginTop="10dp"
+            android:textSize="14sp"
+            android:fontFamily="@font/pretendard_regular"
+            android:text="사용자님의 등록했던 정보를\n확인할 수 있어요"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/tv_home_my_title"/>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
