+++ app/src/main/res/layout/fragment_search.xml
@@ -72,6 +72,20 @@
                 android:background="@drawable/rectangle_gray_8"
                 android:text="20개 더보기"/>
 
+            <TextView
+                android:id="@+id/btn_register_log"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toBottomOf="@id/btn_more"
+                app:layout_constraintEnd_toEndOf="parent"
+                android:elevation="1dp"
+                android:fontFamily="@font/pretendard_regular"
+                android:padding="10dp"
+                android:layout_marginTop="10dp"
+                android:layout_marginEnd="10dp"
+                android:background="@drawable/rectangle_gray_8"
+                android:text="검색 내역 저장"/>
+
             <androidx.coordinatorlayout.widget.CoordinatorLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
